backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
dashscope==1.25.13
et_xmlfile==2.0.0
faiss-cpu==1.13.2
filelock==3.25.1
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
propcache==0.4.1
pycparser==3.0
Pygments==2.19.2
PyMuPDF==1.27.2
PyMySQL==1.1.2
PyPDF2==3.0.1
python-docx==1.2.0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==82.0.1
shellingham==1.5.4
SQLAlchemy==2.0.48
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
Werkzeug==3.1.6
yarl==1.23.0
